requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile
#
asgiref==3.4.1
    # via django
certifi==2021.5.30
    # via requests
cffi==1.14.5
    # via cryptography
chardet==4.0.0
    # via requests
cryptography==3.4.7
    # via pyjwt
defusedxml==0.7.1
    # via python3-openid
dj-database-url==0.5.0
    # via django-heroku
django==3.2.5
    # via
    #   -r requirements.in
    #   django-allauth
    #   django-debug-toolbar
    #   django-filter
    #   django-heroku
django-allauth==0.44.0
    # via -r requirements.in
django-debug-toolbar==3.2.1
    # via -r requirements.in
django-environ==0.4.5
    # via -r requirements.in
django-filter==2.4.0
    # via -r requirements.in
django-heroku==0.3.1
    # via -r requirements.in
gunicorn==20.1.0
    # via -r requirements.in
idna==2.10
    # via requests
oauthlib==3.1.1
    # via requests-oauthlib
psycopg2==2.9.1
    # via django-heroku
psycopg2-binary==2.9.1
    # via -r requirements.in
pycparser==2.20
    # via cffi
pyjwt[crypto]==2.1.0
    # via django-allauth
python3-openid==3.2.0
    # via django-allauth
pytz==2021.1
    # via django
requests==2.25.1
    # via
    #   django-allauth
    #   requests-oauthlib
requests-oauthlib==1.3.0
    # via django-allauth
sqlparse==0.4.1
    # via
    #   django
    #   django-debug-toolbar
urllib3==1.26.6
    # via requests
whitenoise==5.2.0
    # via django-heroku

# The following packages are considered to be unsafe in a requirements file:
# setuptools
